add ut test for the event exporter
